backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib
pydantic-settings
email-validator
alembic
stripe
python-multipart
httpx
pytest
pytest-asyncio
